requirements.txt
# Core FastAPI and GraphQL dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
strawberry-graphql[fastapi]==0.217.0

# Azure SDK dependencies
azure-cosmos==4.5.1
azure-storage-blob==12.19.0
azure-identity==1.15.0

# Configuration and validation
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# HTTP client
httpx==0.25.2
requests==2.32.3
PyJWT==2.10.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Code quality
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Development
python-multipart==0.0.6